Change web scraping wiki page to use rvest
The functions provided by the rvest package are
wrappers to those provided by httr and xml2. It looks like
it is the best way to do web scraping in R generally.

This required quite a bit of the original page to be edited.
It did, however, simplify the whole process being described
so that's a good thing.